Move contact-syncer to add tests
