+++ opengever/contact/syncer/__init__.py
@@ -1,0 +1,7 @@
+from opengever.contact.syncer.object_syncer import OrganizationSyncer  # noqa
+from opengever.contact.syncer.object_syncer import PersonSyncer  # noqa
+from opengever.contact.syncer.object_syncer import MailSyncer  # noqa
+from opengever.contact.syncer.object_syncer import UrlSyncer  # noqa
+from opengever.contact.syncer.object_syncer import AddressSyncer  # noqa
+from opengever.contact.syncer.object_syncer import PhoneNumberSyncer  # noqa
+from opengever.contact.syncer.object_syncer import OrgRoleSyncer  # noqa
